feat: 엔티티 필드 및 클래스 이름 변경
- library -> bookmark 이름 변경
- 채팅방 제목 nullable 허용
- message에 uuid 추가
- 의미 없는 주석 제거
- user 최고 프롬프트 점수 실수형 변환

+++ back/src/main/java/com/closeai/ecoprompt/userinfo/model/entity/UserInfo.java
@@ -20,7 +20,7 @@ public class UserInfo extends BaseEntity {
     private Integer totalMileage;
 
     @Column(name = "high_score", nullable = false)
-    private Integer highScore;
+    private Double highScore;
 
     @Column(name = "personal_prompt", length = 1000)
     private String personalPrompt;
